backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.5.2
    # via django
asttokens==2.0.8
    # via stack-data
attrs==22.1.0
    # via jsonschema
autopep8==1.7.0
    # via -r requirements.in
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.1
    # via nbconvert
certifi==2022.6.15.1
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==2.1.1
    # via requests
colorama==0.4.5
    # via ipython
coverage==6.4.4
    # via -r requirements.in
debugpy==1.6.3
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
django==4.1.1
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.13.0
    # via -r requirements.in
django-dotenv==1.4.2
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
entrypoints==0.4
    # via jupyter-client
executing==1.0.0
    # via stack-data
faker==14.2.0
    # via -r requirements.in
fastjsonschema==2.16.1
    # via nbformat
graphviz==0.20.1
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via nbconvert
ipykernel==6.15.2
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.5.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   notebook
    #   qtconsole
ipywidgets==8.0.2
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   nbconvert
    #   notebook
jsonschema==4.16.0
    # via nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.3.5
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.4
    # via jupyter
jupyter-core==4.11.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==3.0.3
    # via ipywidgets
lxml==4.9.1
    # via nbconvert
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==2.0.4
    # via nbconvert
nbclient==0.6.8
    # via nbconvert
nbconvert==7.0.0
    # via
    #   jupyter
    #   notebook
nbformat==5.4.0
    # via
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.12
    # via jupyter
numpy==1.23.2
    # via pandas
packaging==21.3
    # via
    #   ipykernel
    #   nbconvert
    #   qtpy
pandas==1.4.4
    # via -r requirements.in
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pickleshare==0.7.5
    # via ipython
pillow==9.2.0
    # via -r requirements.in
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.31
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.2
    # via ipykernel
psycopg2-binary==2.9.3
    # via -r requirements.in
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.9.1
    # via autopep8
pycparser==2.21
    # via cffi
pydot==1.4.2
    # via -r requirements.in
pydot-ng==2.0.0
    # via -r requirements.in
pygments==2.13.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyjwt==2.4.0
    # via djangorestframework-simplejwt
pyparsing==3.0.9
    # via
    #   packaging
    #   pydot
    #   pydot-ng
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   faker
    #   jupyter-client
    #   pandas
pytz==2022.2.1
    # via
    #   djangorestframework
    #   pandas
pywin32==304
    # via jupyter-core
pywinpty==2.0.7
    # via terminado
pyzmq==23.2.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.3.2
    # via jupyter
qtpy==2.2.0
    # via qtconsole
requests==2.28.1
    # via -r requirements.in
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
stack-data==0.5.0
    # via ipython
terminado==0.15.0
    # via notebook
tinycss2==1.1.1
    # via nbconvert
toml==0.10.2
    # via autopep8
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.3.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
tzdata==2022.2
    # via django
urllib3==1.26.12
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
widgetsnbextension==4.0.3
    # via ipywidgets
zipp==3.8.1
    # via importlib-metadata
